+++ CHANGES.rst
@@ -214,8 +214,9 @@ astropy.convolution
 astropy.coordinates
 ^^^^^^^^^^^^^^^^^^^
 
-- Catch the error case case of a ``QuantityAttribute`` defaulting to None at
-  attribute creation, and give a more useful error message [#8300]
+- Convert the default of ``QuantityAttribute``, thereby catching the error case
+  case of it being set to None at attribute creation, and giving a more useful
+  error message in the process. [#8300]
 
 astropy.cosmology
 ^^^^^^^^^^^^^^^^^